Move IOBiConsumer into org.primefaces.util
