+++ primefaces/src/main/java/org/primefaces/component/badge/BadgeRenderer.java
@@ -29,7 +29,7 @@ import javax.faces.component.UIComponent;
 import javax.faces.context.FacesContext;
 import javax.faces.context.ResponseWriter;
 
-import org.primefaces.functional.IOBiConsumer;
+import org.primefaces.util.IOBiConsumer;
 import org.primefaces.model.badge.BadgeModel;
 import org.primefaces.renderkit.CoreRenderer;
 import org.primefaces.util.LangUtils;
