+++ primefaces/src/main/java/org/primefaces/util/IOBiConsumer.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.primefaces.functional;
+package org.primefaces.util;
 
 import java.io.IOException;
 import java.util.Objects;
